requirements.txt
coverage==4.5.3
Django==1.11.20
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
